Move one.sample.dev-server to one.dev-server. Refs. #94.
The application configuration is now assumed to always be located
in the one.application namespace. This change allows the dev-server
to become part of the library so that it can now be run from library
code and be used to run library tests.

The change means that we have crossed the line from library to
framework.