Explore messaging and code organization ideas
The single "state" atom contains all essential state for the app. While
this has the advantage of keeping all state in one atom (which
guarantees atomic updates to application state), it has the disadvantage
of causing the atom to hold a sizable amount of data (primarily due to
storing the full list of exercises in the atom). Changes to the state
atom fire a :model-change event, and that event includes the old state
of the atom, the new state of the atom, and a message. This causes the
size of the events to be very large. In testing with ~130 exercises,
this increase in message size significantly degrades performance. The
degradation negatively affects the user experience to such a degree
that this approach is not a viable long-term solution.

Changes:

  - Add more functionality, so that code organization and messaging
    between components is a bit more interesting:
    - Add workflow for initializing the datastore:
      - Prompt the user to enter his/her MongoHQ API key
      - Send a request to MongoHQ to verify that the API key works
      - Send a request to MongoHQ to verify that the database exists
        (since MongoHQ does not permit the REST API to create databases
        for free accounts
      - Send a request to MongoHQ to verify the "exercises" collection
        exists, and if necessary, send a request to MongoHQ to create
        the "exercises" collection
    - Store the API key in a cookie after verifying that it works
    - Display visual feedback in the UI as the app progresses through
      the workflow described above
  - Replace JSONP with XHR. JSONP can only issue GET requests. The
    MongoHQ REST API uses proper HTTP methods (e.g., POST for creating a
    resource, DELETE for deleting a resource), so JSONP won't meet our
    needs.
  - Code organization exploration:
    - Separate view code into distinct namespaces depending on the part
      of the app to which the code pertains (e.g., datastore
      configuration is separate from working with the list of exercises)
    - Move the model-related "after hooks" into distinct namespaces
      depending on the part of the model to which the hooks pertain.
      (If we had agents in ClojureScript, these "after hooks" would
      probably be handled as dispatches that occur within the
      transaction that updates the agent.)
    - Remove the controller