Simplify content of :exercise-search results map
Large :model-change messages cause noticeable performance degradation in
the UI. In an effort to have smaller :model-change messages, we want to
reduce non-essential content from the messages.

Prior to this commit, the :exercise-search map (inside the state atom)
contained a list of all exercises that matched the search. That list
included all exercise attributes (i.e., the ID and the name). Since we
can meet the needs of the application with just the exercise IDs, this
commit changes the logic to store only the exercise IDs in the
:exercise-search map.