Enhance #find-documents to support a 'limit' arg
So far, because the list of exercise documents is fairly small (~150),
and because we want to show all exercises in the exercise list, we have
always fetched _all_ documents whenever querying a MongoHQ collection.

Soon, we want to start fetching the list of _recent_ exercise sets.
Exercises will have a _lot_ of sets, and we don't want the overhead of
fetching _all_ sets when we only want the recent ones. Now that we can
limit the number of documents returned by #find-documents, we are part
of the way toward our goal of fetching a limited number of recent
exercise sets. Next, we'll need to enhance #find-documents to support a
'sort' argument. We'll combine that argument with the 'limit' argument
to allow us to fetch a limited number of recent exercise sets.