Allow user to specify the database name
MongoHQ will not allow two different users to have databases with the
same name. (This *might* not be the case for paid MongoHQ accounts, but
it is the case for the free "sandbox" databases.) Therefore, we cannot
hard-code the database name; instead, we need to prompt the user to
enter his/her unique database name.

  - Update UI to prompt user to enter the database name
  - Inside the :state atom, enhance the :database-configuration to
    include the database name
  - For all functions that formerly took the MongoHQ API key as an
    argument, update those functions to instead take the database
    configuration map as an argument. (This map includes both the API key
    and the database name.)
  - Move clj->js from one.repmax.mongohq to new one.repmax.util
    namespace
  - Introduce one.repmax.url-blueprint