Do not "auto-authenticate" in development mode
For integration testing, we want to test the full end-to-end workflow.
If the browser already has the the user's authentication info stored in
a cookie, we still want the integration tests to fill in the
authentication info. In order to ensure that the integration tests don't
use the authentication info stored in the cookies (which would bypass
the step of entering the authentication info into the UI form fields),
let's try having the app ignore the cookies in development mode. In
production mode, for convenience, the app will continue to use the
authentication info stored in the cookies (if present).