WIP Sketch early ideas for reading auth data from file
The full-stack integration tests need authentication info in order to
contact MongoHQ, but we don't want to store that info in the repo.
Instead, we'd like to store that info in a local file (i.e.,
resources/datastore.clj) that is listed in .gitignore.

This commit is just a sketch of how that might eventually work. We need
to read the file with Clojure, but we need to pass the contents (i.e.,
the API key and the database name) to ClojureScript. Seems like tricky
business, and this commit does not yet provide a working implementation.